@@ -2,57 +2,67 @@
     "name": "student-workflow-api",
     "version": "1.0.0",
     "description": "",
-    "main": "app.js",
+    "main": "app.ts",
     "scripts": {
-        "debug": "ndb server.js",
-        "dev": " NODE_ENV=DEVELOPMENT nodemon server.js",
-        "prod": "node server.js NODE_ENV=PRODUCTION",
-        "start": "nodemon server.js",
+        "debug": "ndb server.ts",
+        "dev": " NODE_ENV=DEVELOPMENT nodemon server.ts",
+        "prod": "node server.ts NODE_ENV=PRODUCTION",
+        "start": "nodemon server.ts",
         "test": "jest --detectOpenHandles",
         "prepare": "cd .. && husky install"
     },
     "author": "Bella Abdelouahab",
     "license": "ISC",
     "dependencies": {
-        "axios": "^1.4.0",
-        "bcryptjs": "^2.4.3",
-        "compression": "^1.7.4",
-        "cookie-parser": "1.4.6",
+        "@typescript-eslint/parser": "6.5.0",
+        "axios": "1.5.0",
+        "bcrypt": "5.1.1",
+        "compression": "1.7.4",
+        "cookie-parser": "^1.4.5",
         "cors": "^2.8.5",
-        "dotenv": "^16.0.3",
-        "express": "^4.18.2",
+        "dotenv": "16.3.1",
+        "express": "^4.17.1",
         "express-bearer-token": "2.4.0",
-        "express-mongo-sanitize": "^2.2.0",
-        "express-rate-limit": "^6.7.0",
+        "express-mongo-sanitize": "2.2.0",
+        "express-rate-limit": "6.10.0",
         "express-routes-versioning": "1.0.1",
-        "helmet": "^6.0.1",
-        "hpp": "^0.2.3",
-        "http-status-codes": "^2.2.0",
-        "jest": "^29.5.0",
-        "jsonwebtoken": "^9.0.0",
-        "mongoose": "^6.8.4",
-        "morgan": "^1.10.0",
-        "qs": "^6.11.2",
-        "supertest": "^6.3.3",
-        "swagger-autogen": "^2.23.4",
-        "swagger-ui-express": "^4.6.3",
-        "validator": "^13.7.0",
-        "winston": "^3.8.2",
+        "helmet": "4.6.0",
+        "hpp": "0.2.3",
+        "http-status-codes": "2.2.0",
+        "jsonwebtoken": "9.0.2",
+        "mongoose": "7.5.0",
+        "morgan": "1.10.0",
+        "qs": "6.11.2",
+        "supertest": "6.3.3",
+        "swagger-autogen": "2.23.5",
+        "swagger-ui-express": "5.0.0",
+        "validator": "^13.11.0",
+        "winston": "^3.10.0",
         "winston-daily-rotate-file": "^4.7.1",
-        "xss-clean": "^0.1.1",
+        "xss-clean": "^0.1.4",
         "yamljs": "^0.3.0"
     },
     "devDependencies": {
-        "eslint": "8.45.0",
-        "eslint-config-prettier": "8.8.0",
+        "@types/body-parser": "^1.19.0",
+        "@types/cookie-parser": "^1.4.2",
+        "@types/cors": "^2.8.10",
+        "@types/express": "^4.17.13",
+        "@types/helmet": "^4.0.0",
+        "@types/node": "^16.11.7",
+        "eslint": "8.48.0",
+        "eslint-config-prettier": "9.0.0",
         "eslint-plugin-prettier": "5.0.0",
-        "husky": "8.0.3",
-        "lint-staged": "13.2.3",
-        "nodemon": "3.0.1",
-        "prettier": "3.0.0"
+        "husky": "7.0.4",
+        "jest": "27.5.1",
+        "lint-staged": "14.0.1",
+        "nodemon": "2.0.22",
+        "prettier": "3.0.3",
+        "ts-jest": "^27.1.3",
+        "ts-node": "^10.4.0",
+        "typescript": "^4.4.4"
     },
     "lint-staged": {
-        "*.+(js)": "eslint --fix",
-        "*.+(js|json)": "prettier --config .prettierrc.json --write ."
+        "*.+(ts|js)": "eslint --fix",
+        "*.+(ts|json|js)": "prettier --config .prettierrc.json --write ."
     }
 }
